Update Function names to be camelCase
Function names from core_jobs.c now in camelCase.
Add --nondet-static flag back to run_proofs.sh
Update static variables in jobs.c so cbmc does not havoc them

+++ test/cbmc/stubs/strnlen.c
@@ -6,7 +6,7 @@ size_t strnlen(const char * s, size_t maxlen)
 
     size_t result;
 
-    result = strlen(s);
+    result = (size_t) strlen(s);
     if (result > maxlen){
         result = maxlen;
     }
